+++ src/Microsoft.Sbom.Api/Workflows/SbomParserBasedValidationWorkflow.cs
@@ -18,8 +18,8 @@ using Microsoft.Sbom.Api.SignValidator;
 using Microsoft.Sbom.Api.Utils;
 using Microsoft.Sbom.Api.Workflows.Helpers;
 using Microsoft.Sbom.Common;
-using Microsoft.Sbom.Common.ComplianceStandard;
 using Microsoft.Sbom.Common.Config;
+using Microsoft.Sbom.Common.ConformanceStandard;
 using Microsoft.Sbom.Extensions;
 using Microsoft.Sbom.JsonAsynchronousNodeKit;
 using Microsoft.Sbom.Parser;
