+++ src/Microsoft.Sbom.Common/ConformanceStandard/Enums/NTIAErrorType.cs
@@ -4,7 +4,7 @@
 using System;
 using Microsoft.Sbom.Parsers.Spdx30SbomParser.ComplianceStandard.Interfaces;
 
-namespace Microsoft.Sbom.Common.ComplianceStandard.Enums;
+namespace Microsoft.Sbom.Common.ConformanceStandard.Enums;
 
 public class NTIAErrorType : IComplianceStandardErrorType, IEquatable<NTIAErrorType>
 {
