+++ src/Microsoft.Sbom.Common/ConformanceStandard/Interfaces/IComplianceStandardEnforcer.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Text.Json;
 using Microsoft.Sbom.Common;
-using Microsoft.Sbom.Common.ComplianceStandard;
+using Microsoft.Sbom.Common.ConformanceStandard;
 using Microsoft.Sbom.Contracts.Enums;
 
 namespace Microsoft.Sbom.Parsers.Spdx30SbomParser.ComplianceStandard.Interfaces;
