+++ src/Microsoft.Sbom.Common/ElementsResult.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using Microsoft.Sbom.Common.ComplianceStandard;
+using Microsoft.Sbom.Common.ConformanceStandard;
 using Microsoft.Sbom.Common.Spdx30Entities;
 using Microsoft.Sbom.JsonAsynchronousNodeKit;
