+++ test/Microsoft.Sbom.Api.Tests/Workflows/SbomParserBasedValidationWorkflowTests.cs
@@ -21,9 +21,9 @@ using Microsoft.Sbom.Api.Utils;
 using Microsoft.Sbom.Api.Workflows;
 using Microsoft.Sbom.Api.Workflows.Helpers;
 using Microsoft.Sbom.Common;
-using Microsoft.Sbom.Common.ComplianceStandard;
-using Microsoft.Sbom.Common.ComplianceStandard.Enums;
 using Microsoft.Sbom.Common.Config;
+using Microsoft.Sbom.Common.ConformanceStandard;
+using Microsoft.Sbom.Common.ConformanceStandard.Enums;
 using Microsoft.Sbom.Contracts;
 using Microsoft.Sbom.Contracts.Enums;
 using Microsoft.Sbom.Extensions;
