+++ test/Microsoft.Sbom.Parsers.Spdx30SbomParser.Tests/Parser/SbomPackageParserTests.cs
@@ -4,7 +4,7 @@
 using System.IO;
 using System.Linq;
 using System.Text;
-using Microsoft.Sbom.Common.ComplianceStandard.Enums;
+using Microsoft.Sbom.Common.ConformanceStandard.Enums;
 using Microsoft.Sbom.JsonAsynchronousNodeKit.Exceptions;
 using Microsoft.Sbom.Parser.JsonStrings;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
